Add minimal release configuration.
- Configure CMake to build and test the library;
- Add pyproject.toml files for runtime-only and development wheels;
- Add GitHub CI for major platforms;
- Add .gitignore;
- Restore moduleDeps.cpp to explicitly register Python module deps;
- Add minimal testWrapper.py script to run Python tests;
- Define DEFFILEMODE if missing to support musl-based ARM toolchains.
- Fix test failures from 'testenv' Python module removal;
- Fix lib-loading tests run from build directory;
- Resolve C4244 warnings in hash.cpp/denseHashMap.cpp tests;
- Fix Windows path in pathUtils.cpp test.

+++ test/hash.cpp
@@ -268,7 +268,7 @@ Test_TfHash()
     std::map<int, uint32_t> mint = {{-1, 1}, {2, 3}, {-4, 5}};
     printf("hash(map<int, uint32_t>): %zu\n", h(mint));
 
-    std::pair<int, float> intfloat = {1, 2.34};
+    std::pair<int, float> intfloat = {1, static_cast<float>(2.34)};
     printf("hash(pair<int, float>): %zu\n", h(intfloat));
 
     std::tuple<int, float> intfloat2 = {1, 2.34};
